Blogging, are we?
The functionalities of the blog module are working. Ony css is needed.

+++ blog/views.py
@@ -1,18 +1,22 @@
-from django.contrib import messages
 from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect
 from .forms import PostForm
 from .models import Post
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.utils import timezone
+from django.http import Http404
+from django.db.models import Q
+
 # Create your views here.
 def post_create(request):
+    if not request.user.is_staff:
+        raise Http404("User not staff")
 
-    form = PostForm(request.POST or None)
+
+    form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance= form.save(commit=False)
         instance.save()
-        messages.success(request, 'Succesfully Created')
         return HttpResponseRedirect(instance.get_absolute_url())
-    else:
-        messages.error(request, 'Succesfully Created')
     context = {
         'form': form
     }
@@ -20,24 +24,48 @@ def post_create(request):
 
 def post_detail(request, pk=None):
     instance = get_object_or_404(Post, id=pk)
-    context= { 'instance' : instance}
+    if instance.draft and not request.user.is_staff:
+        raise Http404('Article not found')
+    context= {
+        'obj' : instance,
+        'staff' : request.user.is_staff
+    }
 
     return render(request, 'blog/post_detail.html', context)
 
 def post_list(request):
-    queryset= Post.objects.all()
+    if request.user.is_staff:
+        queryset= Post.objects.all()
+    else:
+        queryset= Post.objects.filter(draft=False).filter(published__lte = timezone.now()).all()
+
+    query = request.GET.get('q')
+    if query:
+        queryset = queryset.filter(
+        Q(title__icontains=query) |
+        Q(content__icontains=query)
+        ).distinct()
+
+    paginator = Paginator(queryset, 2)
+    page = request.GET.get('page')
+    posts = paginator.get_page(page)
+
     context = {
-        'object_list' : queryset,
+        'object_list' : posts,
+        'staff' : request.user.is_staff
     }
-    return render(request, 'blog/base.html', context)
+    return render(request, 'blog/post_list.html', context)
+
+
 
 def post_update(request, pk=None):
+    if not request.user.is_staff:
+        raise Http404("User not staff")
     instance = get_object_or_404(Post, id=pk)
-    form = PostForm(request.POST or None, instance= instance)
+    form = PostForm(request.POST or None, request.FILES or None, instance= instance)
     if form.is_valid():
         instance= form.save(commit=False)
         instance.save()
-        messages.success(request, 'Succesfully updated')
         return HttpResponseRedirect(instance.get_absolute_url())
 
     context = {
@@ -47,7 +75,8 @@ def post_update(request, pk=None):
     return render(request, 'blog/post_form.html', context)
 
 def post_delete(request, pk=None):
+    if not request.user.is_staff:
+        raise Http404("User not staff")
     instance = get_object_or_404(Post, id=pk)
     instance.delete()
-    messages.success(request, 'Succesfully updated')
     return redirect('bloglist')
